remove the rejection of datasets
Checking by identifier and info rejects also datasets that we want to import.
The reason why this does not work is not clear: it seems that in some cases
the import should be done despite the fact that the harvest source is not added in the extras of the dataset.