Fix logic and clarify what is actually mapped
Resource node urls that have one of the landing_page_protocols may be mapped onto a dataset as its landing page, but only if the dataset does not already have a landing page. In this case, they should not be added as relations in the dataset.

Resource node urls that have one of the additional_relation_protocols should be added to the dataset's relations, but not mapped as its landing page.

The code should now be easier to understand, and also not add urls as relations if they have been mapped as a dataset's url.