test: Check that the correct dataset was deleted
Also correct the expected number of harvest objects: since this test was first written, we have fixed the bug where new datasets were created on every harvest run, even if the data had not changed, instead of already-existing datasets being updated. Thus, we only expect two datasets to be associated with the harvester after the jobs in the test have been run.