fix: bug in mapping of publisher url
the publisher url retrieval is implemented as such that first the
publisher name is derived via a list of roles: 'publisher', 'owner',
etc.
On the role where the publisher name was found, it is attempted to
also get a url. if that fails the default url, which is that of
the organization of the dataset on opendata.swiss is used as
publisher_url