feat: store protocol and normed protocol in resource
protocol is the original protocol that comes from geocat as
gmd:protocol

normend protocol is the derived protocol: this protocol is derived
from the original protocol and decides how the resource is mapped

these two properties are stored on the resource in order to spot
errors in the mapping of resources from geocat more easily

the protocol is besides the description the only field that is
taken as is from a geocat resource. All other fields are transformed
during the mapping: how they are transformed depends on the
normed_protocol. Therefore both protocols are added as resource
fields.

The normed protocol should ideally match the protocol, since the
longer protocol names are deprecated but still allowed and served