fix: resource title mapping
if the gmd:name is not set for a geocat resource the resource
title should be empty

there was a bug in how the resource name was picked up: the path
gmd:name was not used for that: it will be used now

the mapping of the mane of geocat resources to the dataset resources
was simplified: for geocat resources and services the name field is set
if an name was found. This name will then be mapped to the dataset
resources. If no name was found the title will be an empty language
dictionary. Only for map preview the name must be prefixed, so that
the frontend will be able to identify those resources