feat: Make cql configurable
V4 of the geocat csw endpoint will use the cql "subject = 'opendata.swiss'" instead of "keyword = 'opendata.swiss'" to filter for Open Data datasets.
Therefore we need to be able to configure the cql to use per harvester. This config option is already set in all the geocat harvesters, but the values were overwritten.