feat: Adapt CSW processor to v4
There seems to be a bug in the v4 csw-endpoint, where we can get a CatalogServiceWeb.results array like this while harvesting: {'matches': 123, 'nextrecord': 124, 'returned': 23}.
This means that we would next request the 50 records starting with no. 124, which don't exist, so an error is returned.
Adapting the comparison prevents this, but should not cause problems for v3 csw-endpoints.