feat: harvester writes custom activities if a dataset has changed
this features customizes the check on when a dataset has changed
when a change of the dataset is detected the user harvest-notification
creates an activity of type 'package changed'.

Use case:
These activities can be used to trigger notifications in
ckanext-subscribe rather then the automaitcally generated
activities by the user harvest.

Since the activity type is a fixed vocabulary in CKAN 2.8 the filtering
needs to be done by the user who creates the activity rather then by
the activity type.